+++ Tests/Fixtures/config/env_param.php
@@ -2,10 +2,6 @@
 
 use Symfony\Component\DependencyInjection\Tests\Fixtures\AcmeConfig;
 
-return function (AcmeConfig $config, string $env) {
-    if ('prod' === $env) {
-        $config->color('blue');
-    } else {
-        $config->color('red');
-    }
+return function (string $env) {
+    return new AcmeConfig(['color' => 'prod' === $env ? 'blue' : 'red']);
 };