+++ Tests/Fixtures/config/return_config_builder.php
@@ -2,7 +2,4 @@
 
 use Symfony\Component\DependencyInjection\Tests\Fixtures\AcmeConfig;
 
-$config = new AcmeConfig();
-$config->color('red');
-
-return $config;
+return new AcmeConfig(['color' => 'red']);